refactor(wasm): split space_plate's JsValue-free core into its own module
CI's Rust module-size ratchet failed: extracting the pure functions for
direct testing pushed space_plate.rs to 408 lines, past the 400-line
threshold with no allowlist row.

The ratchet's rule is SPLIT, never add a row, so this splits rather than
allowlists. space_plate_input.rs now holds the JsValue-free core — the DoS
ceilings, resolve_build_options, build_segments, build_wall_rects and the
EditError -> stable-code table — and space_plate.rs keeps the
#[wasm_bindgen] entry points as thin wrappers. 408 -> 325 lines.

Two knock-ons, both shrunk rather than raised:
- clippy flagged BuildOptions and InputSegment as unused in space_plate.rs
  once they moved out; dropped from its import list.
- mod.rs went 1092 -> 1093 from the new  line, one past its budget.
  Folded the three-line module doc into one line: 1091, under budget.

module_size_ratchet 5/5, ifc-lite-wasm 124 tests passing, clippy
-D warnings clean. The gate also now notes grid_lines.rs has dropped to
327 and its allowlist row can be removed — left for a separate change.
