+++ rust/wasm-bindings/src/api/mod.rs
@@ -2,9 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! JavaScript API for IFC-Lite
-//!
-//! Modern async/await API for parsing IFC files.
+//! JavaScript API for IFC-Lite: modern async/await parsing.
 
 mod alignment_lines;
 mod bool2d;
@@ -26,6 +24,7 @@ mod parsing;
 mod pipeline_diagnostics;
 mod simplify;
 mod space_plate;
+mod space_plate_input;
 pub(crate) mod styling;
 mod symbolic;
 mod zone_split;
